run: fix log paths in results.json
The "output" links from NIPA's contest page [1] points to a directory
in "raw.githubusercontent.com", which is not working, e.g. [2].

The result.json file is already modified to point to 'github.com'
instead, but for a bit of time now, there is a second 'link' field in
the JSON file, e.g. [3]. It needs to be modified as well.

Instead of converting the fields one by one, one 'sed' command can
replace all references to 'githubusercontent.com'.

Link: https://netdev.bots.linux.dev/contest.html?remote=tdc [1]
Link: https://raw.githubusercontent.com/p4tc-dev/tc-executor/storage/artifacts/643403/1-tdc-sh [2]
Link: https://raw.githubusercontent.com/p4tc-dev/tc-executor/storage/results/results-636563.json [3]